+++ app/api/pgs.py
@@ -98,6 +98,31 @@ class Pgs(Restful):
             self.logger.error(f"StatusCode:{trx.rsp_status_code} {trx.rsp_text}")
         return trx
 
+    def get_clientHandle(self, trx, tokenReq):
+        trx.shop = self._getShop()
+        trx.shopInfo = self._getShopInfo()
+        featureURL = "/tokenization/{tenant}".format(tenant=self._getTenant())
+        body = {"requestor": f"{trx.shop}", 
+                # "id": "1",
+                "correlationId": trx.correlationId,
+                "locale": self._getLocale(),
+                "successCallbackUrl": f"{self._getSuccessUrl()}",
+                "failureCallbackUrl": f"{self._getFailureUrl()}",
+                "customStyle": "style=\"color:red;\""
+                }
+
+        self.logger.debug(featureURL + " " + json.dumps(body))
+        resp = self.put(self._url(featureURL), 
+                        data=json.dumps(body), 
+                        auth=self.auth)
+        trx.rsp_status_code = resp.status_code
+        trx.rsp_text = resp.text
+        if trx.rsp_status_code == 401:
+            self.logger.warn(f"StatusCode:{trx.rsp_status_code} {json.dumps(json.loads(trx.rsp_text))}")
+        else:
+            self.logger.error(f"StatusCode:{trx.rsp_status_code} {trx.rsp_text}")
+        return trx
+
     def pay_token_cart(self, trx, requester, correlation, cart, token, initiator):
         featureURL = "/paymentwithtoken/{tenant}".format(tenant=self._getTenant())
         body = {"requestor": f"{requester}", 
@@ -131,6 +156,33 @@ class Pgs(Restful):
             locale=self._getLocale(),
             costCentre=trx.costCentre,
             imageColor=trx.imageColor,
+            )
+
+        self.logger.debug(featureURL)
+        resp = self.get(self._url(featureURL), 
+                        data=None, 
+                        auth=self.auth)
+        trx.rsp_status_code = resp.status_code
+        trx.rsp_text = resp.text
+        if trx.rsp_status_code == 401:
+            self.logger.warn(f"StatusCode:{trx.rsp_status_code} {json.dumps(json.loads(trx.rsp_text))}")
+        elif trx.rsp_status_code == 500:
+            self.logger.error(f"StatusCode:{trx.rsp_status_code} {trx.rsp_text}")
+        else:
+            self.logger.debug(f"StatusCode:{trx.rsp_status_code}")
+        return trx
+
+    def get_tokenization_methods(self, trx):
+        trx.costCentre = self._getCostCentre()
+        trx.imageColor = self._getImageColor()
+        featureURL = "/tokenizationtypes/{tenant}/{clientHandle}/{correlationId}/{requestor}/{locale}?costCenter={costCentre}&imageColor={imageColor}".format(
+            tenant=self._getTenant(),
+            clientHandle=trx.clientHandleUuid,
+            correlationId=trx.correlationId,
+            requestor=trx.shop,
+            locale=self._getLocale(),
+            costCentre=trx.costCentre,
+            imageColor=trx.imageColor
             )
 
         self.logger.debug(featureURL)