+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - added Pay & Tokenize feature
 - added Pay with Token feature (last generated token is used for payment)
+- added Tokenize feature (just for 0,00 amount)
 ### Changed
 
 ## [0.9.0](https://github.com/udigeri/inout/compare/v0.8.0..v0.9.0) (2021-12-01)